Very rough attack animation on player.
This is the definitioin of a "proof of concept". It reveals
to me that we definitely need a state machine on the player ASAP.
Without it, the player logic is super weird in the movement system.

It also reveals that we really need split animations/skeletons so
we can have the character walk while moving etc.