Fix drift
When character is in freefall there was drift between the model
and the actual location. This was apprently due to order of operations
and also the lack of syncing the locations within the early return
blocks.

The order of operations is this in the game update loop:

1. Update Physics
2. Update Simulation
3. Update Animations

Otherwise I guess animations are being mathed with stale transform
data and this drifts over time. I don't fullly understand why but
it fixed it so that works.